medical_xai/requirements.txt
streamlit==1.35.0
torch==2.2.0
torchvision==0.17.0
Pillow==10.3.0
numpy==1.26.4
matplotlib==3.8.4
scikit-learn==1.4.2
lime==0.2.0.1
shap==0.45.0
opencv-python==4.9.0.80
pandas==2.2.2
seaborn==0.13.2
plotly==5.22.0
captum==0.7.0
tqdm==4.66.4
scipy==1.13.0
reportlab==4.1.0
